+++ src/minecraft/com/vbitz/MinecraftScript/MinecraftScriptClassShutter.java
@@ -11,6 +11,7 @@ public class MinecraftScriptClassShutter implements ClassShutter {
 	static {
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptPlayerAPI");
+		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptSurvivalPlayerAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptWorldAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptItemStackAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptScriptedBlockAPI");